Authentication working! TODO: writing a lyircs helper

+++ app/controllers/notes_controller.rb
@@ -1,4 +1,6 @@
 class NotesController < ApplicationController
+  skip_before_filter :authorized, :only => [:destroy]
+
   def create
     note = Note.create!(params[:note])
     redirect_to track_url(note.track_id)
@@ -6,9 +8,16 @@ class NotesController < ApplicationController
 
   def destroy
     note = Note.find_by_id(params[:id])
-    track = note.track
-    note.destroy
 
-    redirect_to track_url(track)
+    unless @current_user.is_admin || current_user.id == note.user_id
+      flash[:notices] ||= []
+      flash[:notices] << "You aren't allowed to delete that."
+      redirect_to :back
+    else
+      track = note.track
+      note.destroy
+      redirect_to track_url(track)
+    end
+
   end
 end
